Move PhpSpec to separate namespace
